Update references of db.delegates.insertFork with Account.insertFork
