+++ storage/entities/account.js
@@ -335,6 +335,7 @@ class Account extends BaseEntity {
 			countDuplicatedDelegates: this.adapter.loadSQLFile(
 				'accounts/count_duplicated_delegates.sql'
 			),
+			insertFork: this.adapter.loadSQLFile('accounts/insert_fork.sql'),
 		};
 	}
 
@@ -679,6 +680,29 @@ class Account extends BaseEntity {
 				tx
 			)
 			.then(result => +result.count);
+	}
+
+	// TODO: Should create a separate entity to manage forks
+	/**
+	 * Inserts a fork data table entry.
+	 *
+	 * @param {Object} fork
+	 * @param {string} fork.delegatePublicKey
+	 * @param {integer} fork.blockTimestamp
+	 * @param {string} fork.blockId
+	 * @param {integer} fork.blockHeight
+	 * @param {string} fork.previousBlockId
+	 * @param {string} fork.cause
+	 * @param {Object} [tx] - Database transaction
+	 * @returns {Promise}
+	 */
+	insertFork(fork, tx) {
+		return this.adapter.executeFile(
+			this.SQLs.insertFork,
+			fork,
+			{ expectedResultCount: 0 },
+			tx
+		);
 	}
 
 	/**
